feat: redesign mail queue implementation using sender centric flow
